chore: move event details component
