+++ apps/course-builder-web/src/app/(content)/events/[slug]/_components/event-details.tsx
